chg 将测试代码提出到 activity 中

+++ demo/src/main/java/com/sothree/slidinguppanel/demo/TitleBean.java
@@ -134,7 +134,7 @@ public class TitleBean extends ARefreshable {
         priceArtistSetting.append(String.format("人均:¥ %s  %s", price, typeName)).absoluteSize(SP14);
         
         priceArtistSetting.setMaxLineCount(1);
-        priceArtistSetting.setOuterWidth((int) width);
+        priceArtistSetting.setOuterWidth(width);
         return priceArtistSetting;
     }
